+++ docs/CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 0.1-dev (unreleased)
 --------------------
 
+- Add ability to initialise modem and account with PIN.
+  [davidjb]
 - Add ability for send credit script to detect modem phone number.
   [davidjb]
 - Send credit script does not save data by default (runs unless told otherwise).